migrate tool/agent stack to runenv v2 architecture
Replace the `&dyn RunEnv` trait-object passing model with the v2
`Container`/`Console` split: `ToolFunc::call` now takes an owned
`Arc<RunEnvHandle>`, streams are `'static`, and the HRTB disappears.
`AgentState::runenv` becomes the cloneable v2 `RunEnv` struct; the
agent boots lazily inside `execute_tool_calls` so the VM only starts
when a tool actually runs, and `RunEnv` caches the booted handle so
parallel tool calls share one VM.

The `SharedSandbox` newtype is removed — sharing a sandbox between
agents now collapses to `RunEnv::clone()`. Sandbox-feature tests are
rewritten to use `RunEnv::sandbox(...)` and read back through
`.get().await`.
