+++ src/agent/builder.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use crate::{
     agent::{Agent, AgentProvider, AgentSpec, ContextManager},
     message::Message,
@@ -51,7 +49,7 @@ pub struct AgentBuilder {
 
     history: Vec<Message>,
 
-    runenv: Option<Arc<dyn RunEnv>>,
+    runenv: Option<RunEnv>,
 
     context_manager: Option<ContextManager>,
 }
@@ -126,11 +124,11 @@ impl AgentBuilder {
         self
     }
 
-    /// Use this [`RunEnv`] for tool execution instead of the default [`Local`](crate::runenv::Local).
-    /// Taken by value; wrap an existing [`Arc<Sandbox>`](crate::runenv::Sandbox) in a
-    /// `Clone + RunEnv` newtype if you need to share the same VM across agents.
-    pub fn runenv(mut self, runenv: impl RunEnv) -> Self {
-        self.runenv = Some(Arc::new(runenv));
+    /// Use this [`RunEnv`] for tool execution instead of the default local runenv.
+    /// `RunEnv` is cheaply cloneable, so the same underlying VM can be shared between
+    /// multiple agents by cloning the value.
+    pub fn runenv(mut self, runenv: RunEnv) -> Self {
+        self.runenv = Some(runenv);
         self
     }
 
@@ -176,7 +174,7 @@ impl AgentBuilder {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{agent::AgentCard, lang_model::LangModelProvider, message::Role, runenv::Local};
+    use crate::{agent::AgentCard, lang_model::LangModelProvider, message::Role};
 
     const TEST_MODEL: &str = "openai/gpt-4o-mini";
 
@@ -215,13 +213,12 @@ mod tests {
     async fn test_builder_runenv_is_applied() {
         let agent = AgentBuilder::new(TEST_MODEL)
             .provider(dummy_provider())
-            .runenv(Local {})
+            .runenv(RunEnv::local())
             .build()
             .unwrap();
         // Smoke check: runenv is plugged in and usable.
-        let result = agent
-            .state
-            .runenv
+        let handle = agent.state.runenv.get().await.expect("runenv boot failed");
+        let result = handle
             .exec("sh".into(), vec!["-c".into(), "echo ok".into()], None)
             .await
             .expect("exec failed");
@@ -247,68 +244,34 @@ mod tests {
             .unwrap();
     }
 
-    /// Newtype over `Arc<Sandbox>` so the same underlying VM can be handed to
-    /// multiple builders via `.runenv()`.  `AgentBuilder::runenv` takes
-    /// `impl RunEnv + 'static` by value, so a bare `Arc<Sandbox>` would need
-    /// `RunEnv` to be implemented for `Arc<T>`; this wrapper sidesteps that.
-    #[cfg(feature = "sandbox")]
-    #[derive(Clone)]
-    struct SharedSandbox(Arc<crate::runenv::Sandbox>);
-
-    #[cfg(feature = "sandbox")]
-    #[async_trait::async_trait]
-    impl RunEnv for SharedSandbox {
-        fn get_os(&self) -> &str {
-            "linux"
-        }
-
-        async fn exec(
-            &self,
-            program: String,
-            args: Vec<String>,
-            timeout: Option<u64>,
-        ) -> anyhow::Result<crate::runenv::ExecResult> {
-            self.0.exec(program, args, timeout).await
-        }
-
-        async fn read(&self, path: &std::path::Path) -> anyhow::Result<Vec<u8>> {
-            self.0.read(path).await
-        }
-
-        async fn write(&self, path: &std::path::Path, content: &[u8]) -> anyhow::Result<()> {
-            self.0.write(path, content).await
-        }
-    }
-
-    /// Two agents built with the same `Arc<Sandbox>` (via `SharedSandbox`) see
-    /// each other's filesystem writes — confirms `.runenv()` carries the VM
-    /// reference through to the agent's `state.runenv`.
+    /// Two agents built with the same cloned `RunEnv` see each other's
+    /// filesystem writes — confirms `.runenv()` carries the VM reference
+    /// through to the agent's `state.runenv`, and that cloning the v2
+    /// `RunEnv` shares the underlying container.
     #[cfg(feature = "sandbox")]
     #[tokio::test]
     async fn test_builder_shared_arc_sandbox() {
-        use crate::runenv::{Sandbox, SandboxConfig};
+        use crate::runenv::SandboxConfig;
 
-        let sandbox = Arc::new(
-            Sandbox::new(SandboxConfig::default())
-                .await
-                .expect("sandbox creation failed"),
-        );
-        let shared = SharedSandbox(sandbox.clone());
+        let runenv = RunEnv::sandbox(SandboxConfig::default())
+            .await
+            .expect("sandbox creation failed");
 
         let sub_agent = AgentBuilder::new(TEST_MODEL)
             .provider(dummy_provider())
-            .runenv(shared.clone())
+            .runenv(runenv.clone())
             .build()
             .unwrap();
 
         let parent = AgentBuilder::new(TEST_MODEL)
             .provider(dummy_provider())
-            .runenv(shared.clone())
+            .runenv(runenv.clone())
             .build()
             .unwrap();
 
         // Write through the underlying VM, read back through parent's runenv.
-        sandbox
+        let handle = runenv.get().await.expect("runenv boot failed");
+        handle
             .write(
                 std::path::Path::new("/workspace/shared_test.txt"),
                 b"shared_ok",
@@ -319,6 +282,9 @@ mod tests {
         let bytes = parent
             .state
             .runenv
+            .get()
+            .await
+            .expect("runenv boot failed")
             .read(std::path::Path::new("/workspace/shared_test.txt"))
             .await
             .expect("read failed");
@@ -332,6 +298,9 @@ mod tests {
         parent
             .state
             .runenv
+            .get()
+            .await
+            .expect("runenv boot failed")
             .write(std::path::Path::new("/workspace/shared.txt"), b"shared_ok")
             .await
             .expect("write failed");
@@ -339,6 +308,9 @@ mod tests {
         let bytes = sub_agent
             .state
             .runenv
+            .get()
+            .await
+            .expect("runenv boot failed")
             .read(std::path::Path::new("/workspace/shared.txt"))
             .await
             .expect("subagent runenv should see file written by parent");
