+++ src/lib.rs
@@ -7,6 +7,5 @@ pub mod lang_model;
 pub(crate) mod macros;
 pub mod message;
 pub mod runenv;
-pub mod runenv_v2;
 pub mod tool;
 pub(crate) mod util;
