+++ src/tool/func.rs
@@ -9,19 +9,16 @@ use futures::{
 use crate::{
     datatype::Value,
     message::{FinishReason, Message, MessageOutput, Part, Role},
-    runenv::RunEnv,
+    runenv::RunEnvHandle,
 };
 
 #[derive(Clone)]
 pub struct ToolFunc {
-    // HRTB (`for<'a>`) is required: an async tool that borrows `&'a dyn RunEnv`
-    // produces a future whose lifetime is `'a`, so the resulting stream is
-    // `BoxStream<'a, _>`. Pinning the inner closure to a single concrete
-    // lifetime (e.g. `'static`) would forbid those tools — borrowed `runenv`
-    // could not survive across `.await` points. The HRTB lets the same
-    // `ToolFunc` accept any caller lifetime and propagate it into the stream.
+    // The runenv is passed in as `Arc<RunEnvHandle>`, owned, so the resulting
+    // stream is always `'static`. Tools that need to hold the env across await
+    // points clone the Arc into their stream/future captures.
     inner: Arc<
-        dyn for<'a> Fn(Value, String, &'a dyn RunEnv) -> BoxStream<'a, MessageOutput> + Send + Sync,
+        dyn Fn(Value, String, Arc<RunEnvHandle>) -> BoxStream<'static, MessageOutput> + Send + Sync,
     >,
 }
 
@@ -32,7 +29,7 @@ impl ToolFunc {
     /// only for advanced usage that the macro does not cover.
     pub fn new<F>(f: F) -> Self
     where
-        F: for<'a> Fn(Value, String, &'a dyn RunEnv) -> BoxStream<'a, MessageOutput>
+        F: Fn(Value, String, Arc<RunEnvHandle>) -> BoxStream<'static, MessageOutput>
             + Send
             + Sync
             + 'static,
@@ -40,19 +37,19 @@ impl ToolFunc {
         Self { inner: Arc::new(f) }
     }
 
-    pub fn call<'a>(
+    pub fn call(
         &self,
         args: Value,
         id: impl Into<String>,
-        runenv: &'a dyn RunEnv,
-    ) -> BoxStream<'a, MessageOutput> {
+        runenv: Arc<RunEnvHandle>,
+    ) -> BoxStream<'static, MessageOutput> {
         (self.inner)(args, id.into(), runenv)
     }
 }
 
 /// Wrapping helpers used by the [`crate::tool_func!`] macro to turn user output
 /// (a [`Value`]/[`Message`], an async future of one, or a stream of them) into
-/// the canonical `BoxStream<'a, MessageOutput>` shape.
+/// the canonical `BoxStream<'static, MessageOutput>` shape.
 ///
 /// Not part of the public API.
 #[doc(hidden)]
@@ -81,11 +78,11 @@ pub mod __private {
         .boxed()
     }
 
-    pub fn value_fut_to_stream<'a>(
+    pub fn value_fut_to_stream(
         id: String,
-        fut: BoxFuture<'a, Value>,
-    ) -> BoxStream<'a, MessageOutput> {
-        stream::once(Box::pin(async move {
+        fut: BoxFuture<'static, Value>,
+    ) -> BoxStream<'static, MessageOutput> {
+        stream::once(async move {
             MessageOutput {
                 depth: None,
                 message: Message::new(Role::Tool)
@@ -94,26 +91,28 @@ pub mod __private {
                 finish_reason: FinishReason::Stop {},
                 usage: None,
             }
-        }))
+        })
         .boxed()
     }
 
-    pub fn message_fut_to_stream<'a>(fut: BoxFuture<'a, Message>) -> BoxStream<'a, MessageOutput> {
-        stream::once(Box::pin(async move {
+    pub fn message_fut_to_stream(
+        fut: BoxFuture<'static, Message>,
+    ) -> BoxStream<'static, MessageOutput> {
+        stream::once(async move {
             MessageOutput {
                 depth: None,
                 message: fut.await,
                 finish_reason: FinishReason::Stop {},
                 usage: None,
             }
-        }))
+        })
         .boxed()
     }
 
-    pub fn value_stream_to_msg_stream<'a>(
+    pub fn value_stream_to_msg_stream(
         id: String,
-        s: BoxStream<'a, Value>,
-    ) -> BoxStream<'a, MessageOutput> {
+        s: BoxStream<'static, Value>,
+    ) -> BoxStream<'static, MessageOutput> {
         s.map(move |value| MessageOutput {
             depth: None,
             message: Message::new(Role::Tool)
@@ -125,9 +124,9 @@ pub mod __private {
         .boxed()
     }
 
-    pub fn message_stream_to_msg_stream<'a>(
-        s: BoxStream<'a, Message>,
-    ) -> BoxStream<'a, MessageOutput> {
+    pub fn message_stream_to_msg_stream(
+        s: BoxStream<'static, Message>,
+    ) -> BoxStream<'static, MessageOutput> {
         s.map(|message| MessageOutput {
             depth: None,
             message,
@@ -139,7 +138,7 @@ pub mod __private {
 }
 
 /// Build a [`ToolFunc`] from a closure without writing the canonical
-/// `for<'a> Fn(Value, String, &'a dyn RunEnv) -> BoxStream<'a, MessageOutput>`
+/// `Fn(Value, String, Arc<RunEnvHandle>) -> BoxStream<'static, MessageOutput>`
 /// shape or its `BoxFuture`/`BoxStream`/`Box::pin` boilerplate by hand.
 ///
 /// # Forms
@@ -148,27 +147,27 @@ pub mod __private {
 ///
 /// ```ignore
 /// tool_func!(|args: Value| -> Value { … })
-/// tool_func!(|args: Value, runenv: &dyn RunEnv| -> Value { … })
+/// tool_func!(|args: Value, runenv: Arc<RunEnvHandle>| -> Value { … })
 /// tool_func!(|args: Value, id: String| -> Message { … })
-/// tool_func!(|args: Value, id: String, runenv: &dyn RunEnv| -> Message { … })
+/// tool_func!(|args: Value, id: String, runenv: Arc<RunEnvHandle>| -> Message { … })
 /// ```
 ///
 /// Async — closure body is `async`-awaitable:
 ///
 /// ```ignore
 /// tool_func!(async |args: Value| -> Value { … })
-/// tool_func!(async |args: Value, runenv: &dyn RunEnv| -> Value { … })
+/// tool_func!(async |args: Value, runenv: Arc<RunEnvHandle>| -> Value { … })
 /// tool_func!(async |args: Value, id: String| -> Message { … })
-/// tool_func!(async |args: Value, id: String, runenv: &dyn RunEnv| -> Message { … })
+/// tool_func!(async |args: Value, id: String, runenv: Arc<RunEnvHandle>| -> Message { … })
 /// ```
 ///
 /// Stream — closure body produces a `Stream`/`BoxStream`:
 ///
 /// ```ignore
 /// tool_func!(stream |args: Value| -> Value { … })
-/// tool_func!(stream |args: Value, runenv: &dyn RunEnv| -> Value { … })
+/// tool_func!(stream |args: Value, runenv: Arc<RunEnvHandle>| -> Value { … })
 /// tool_func!(stream |args: Value, id: String| -> Message { … })
-/// tool_func!(stream |args: Value, id: String, runenv: &dyn RunEnv| -> Message { … })
+/// tool_func!(stream |args: Value, id: String, runenv: Arc<RunEnvHandle>| -> Message { … })
 /// ```
 #[macro_export]
 macro_rules! tool_func {
@@ -177,18 +176,18 @@ macro_rules! tool_func {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   id: ::std::string::String,
-                  _runenv: &dyn $crate::runenv::RunEnv| {
+                  _runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 let value: $crate::datatype::Value = $body;
                 $crate::tool::__private::value_to_stream(id, value)
             },
         )
     };
 
-    (|$args:ident : Value, $runenv:ident : &dyn RunEnv| -> Value $body:block) => {
+    (|$args:ident : Value, $runenv:ident : Arc<RunEnvHandle>| -> Value $body:block) => {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   id: ::std::string::String,
-                  $runenv: &dyn $crate::runenv::RunEnv| {
+                  $runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 let value: $crate::datatype::Value = $body;
                 $crate::tool::__private::value_to_stream(id, value)
             },
@@ -199,7 +198,7 @@ macro_rules! tool_func {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   $id: ::std::string::String,
-                  _runenv: &dyn $crate::runenv::RunEnv| {
+                  _runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 let message: $crate::message::Message = $body;
                 $crate::tool::__private::message_to_stream(message)
             },
@@ -207,13 +206,13 @@ macro_rules! tool_func {
     };
 
     (
-        |$args:ident : Value, $id:ident : String, $runenv:ident : &dyn RunEnv| -> Message
+        |$args:ident : Value, $id:ident : String, $runenv:ident : Arc<RunEnvHandle>| -> Message
             $body:block
     ) => {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   $id: ::std::string::String,
-                  $runenv: &dyn $crate::runenv::RunEnv| {
+                  $runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 let message: $crate::message::Message = $body;
                 $crate::tool::__private::message_to_stream(message)
             },
@@ -230,7 +229,7 @@ macro_rules! tool_func {
     // `with [name = expr, ...]` clause to lift those clones up:
     //
     // ```ignore
-    // tool_func!(async |args: Value, runenv: &dyn RunEnv| -> Value
+    // tool_func!(async |args: Value, runenv: Arc<RunEnvHandle>| -> Value
     //     with [runner = runner.clone()]
     //     {
     //         // body sees local `runner` (the per-call clone)
@@ -245,7 +244,7 @@ macro_rules! tool_func {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   id: ::std::string::String,
-                  _runenv: &dyn $crate::runenv::RunEnv| {
+                  _runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 $(let $cap = $expr;)*
                 let fut: ::futures::future::BoxFuture<'static, $crate::datatype::Value> =
                     ::std::boxed::Box::pin(async move $body);
@@ -254,22 +253,22 @@ macro_rules! tool_func {
         )
     };
 
-    (async |$args:ident : Value, $runenv:ident : &dyn RunEnv| -> Value $body:block) => {
+    (async |$args:ident : Value, $runenv:ident : Arc<RunEnvHandle>| -> Value $body:block) => {
         $crate::tool_func!(
-            async |$args: Value, $runenv: &dyn RunEnv| -> Value with [] $body
+            async |$args: Value, $runenv: Arc<RunEnvHandle>| -> Value with [] $body
         )
     };
 
     (
-        async |$args:ident : Value, $runenv:ident : &dyn RunEnv| -> Value
+        async |$args:ident : Value, $runenv:ident : Arc<RunEnvHandle>| -> Value
             with [$($cap:ident = $expr:expr),* $(,)?] $body:block
     ) => {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   id: ::std::string::String,
-                  $runenv: &dyn $crate::runenv::RunEnv| {
+                  $runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 $(let $cap = $expr;)*
-                let fut: ::futures::future::BoxFuture<'_, $crate::datatype::Value> =
+                let fut: ::futures::future::BoxFuture<'static, $crate::datatype::Value> =
                     ::std::boxed::Box::pin(async move $body);
                 $crate::tool::__private::value_fut_to_stream(id, fut)
             },
@@ -289,7 +288,7 @@ macro_rules! tool_func {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   $id: ::std::string::String,
-                  _runenv: &dyn $crate::runenv::RunEnv| {
+                  _runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 $(let $cap = $expr;)*
                 let fut: ::futures::future::BoxFuture<'static, $crate::message::Message> =
                     ::std::boxed::Box::pin(async move $body);
@@ -299,24 +298,24 @@ macro_rules! tool_func {
     };
 
     (
-        async |$args:ident : Value, $id:ident : String, $runenv:ident : &dyn RunEnv|
+        async |$args:ident : Value, $id:ident : String, $runenv:ident : Arc<RunEnvHandle>|
             -> Message $body:block
     ) => {
         $crate::tool_func!(
-            async |$args: Value, $id: String, $runenv: &dyn RunEnv| -> Message with [] $body
+            async |$args: Value, $id: String, $runenv: Arc<RunEnvHandle>| -> Message with [] $body
         )
     };
 
     (
-        async |$args:ident : Value, $id:ident : String, $runenv:ident : &dyn RunEnv|
+        async |$args:ident : Value, $id:ident : String, $runenv:ident : Arc<RunEnvHandle>|
             -> Message with [$($cap:ident = $expr:expr),* $(,)?] $body:block
     ) => {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   $id: ::std::string::String,
-                  $runenv: &dyn $crate::runenv::RunEnv| {
+                  $runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 $(let $cap = $expr;)*
-                let fut: ::futures::future::BoxFuture<'_, $crate::message::Message> =
+                let fut: ::futures::future::BoxFuture<'static, $crate::message::Message> =
                     ::std::boxed::Box::pin(async move $body);
                 $crate::tool::__private::message_fut_to_stream(fut)
             },
@@ -328,7 +327,7 @@ macro_rules! tool_func {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   id: ::std::string::String,
-                  _runenv: &dyn $crate::runenv::RunEnv| {
+                  _runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 let s: ::futures::stream::BoxStream<'static, $crate::datatype::Value> =
                     ::futures::StreamExt::boxed($body);
                 $crate::tool::__private::value_stream_to_msg_stream(id, s)
@@ -336,12 +335,12 @@ macro_rules! tool_func {
         )
     };
 
-    (stream |$args:ident : Value, $runenv:ident : &dyn RunEnv| -> Value $body:block) => {
+    (stream |$args:ident : Value, $runenv:ident : Arc<RunEnvHandle>| -> Value $body:block) => {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   id: ::std::string::String,
-                  $runenv: &dyn $crate::runenv::RunEnv| {
-                let s: ::futures::stream::BoxStream<'_, $crate::datatype::Value> =
+                  $runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
+                let s: ::futures::stream::BoxStream<'static, $crate::datatype::Value> =
                     ::futures::StreamExt::boxed($body);
                 $crate::tool::__private::value_stream_to_msg_stream(id, s)
             },
@@ -352,7 +351,7 @@ macro_rules! tool_func {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   $id: ::std::string::String,
-                  _runenv: &dyn $crate::runenv::RunEnv| {
+                  _runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
                 let s: ::futures::stream::BoxStream<'static, $crate::message::Message> =
                     ::futures::StreamExt::boxed($body);
                 $crate::tool::__private::message_stream_to_msg_stream(s)
@@ -361,14 +360,14 @@ macro_rules! tool_func {
     };
 
     (
-        stream |$args:ident : Value, $id:ident : String, $runenv:ident : &dyn RunEnv|
+        stream |$args:ident : Value, $id:ident : String, $runenv:ident : Arc<RunEnvHandle>|
             -> Message $body:block
     ) => {
         $crate::tool::ToolFunc::new(
             move |$args: $crate::datatype::Value,
                   $id: ::std::string::String,
-                  $runenv: &dyn $crate::runenv::RunEnv| {
-                let s: ::futures::stream::BoxStream<'_, $crate::message::Message> =
+                  $runenv: ::std::sync::Arc<$crate::runenv::RunEnvHandle>| {
+                let s: ::futures::stream::BoxStream<'static, $crate::message::Message> =
                     ::futures::StreamExt::boxed($body);
                 $crate::tool::__private::message_stream_to_msg_stream(s)
             },
@@ -379,14 +378,18 @@ macro_rules! tool_func {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::runenv::Local;
+    use crate::runenv::RunEnv;
+
+    async fn local_handle() -> Arc<RunEnvHandle> {
+        RunEnv::local().get().await.unwrap()
+    }
 
     #[tokio::test]
     async fn test_sync_value() {
-        let runenv = Local {};
+        let runenv = local_handle().await;
         let f = tool_func!(|_args: Value| -> Value { Value::string("ok") });
         let out = f
-            .call(Value::object_empty(), "call-1", &runenv)
+            .call(Value::object_empty(), "call-1", runenv)
             .next()
             .await
             .unwrap();
@@ -399,10 +402,11 @@ mod tests {
 
     #[tokio::test]
     async fn test_sync_value_with_runenv() {
-        let runenv = Local {};
-        let f = tool_func!(|_args: Value, _runenv: &dyn RunEnv| -> Value { Value::string("ok") });
+        let runenv = local_handle().await;
+        let f =
+            tool_func!(|_args: Value, _runenv: Arc<RunEnvHandle>| -> Value { Value::string("ok") });
         let out = f
-            .call(Value::object_empty(), "call-1", &runenv)
+            .call(Value::object_empty(), "call-1", runenv)
             .next()
             .await
             .unwrap();
@@ -414,10 +418,10 @@ mod tests {
 
     #[tokio::test]
     async fn test_async_value() {
-        let runenv = Local {};
+        let runenv = local_handle().await;
         let f = tool_func!(async |_args: Value| -> Value { Value::bool(true) });
         let out = f
-            .call(Value::object_empty(), "call-1", &runenv)
+            .call(Value::object_empty(), "call-1", runenv)
             .next()
             .await
             .unwrap();
@@ -426,11 +430,12 @@ mod tests {
 
     #[tokio::test]
     async fn test_async_value_with_runenv() {
-        let runenv = Local {};
-        let f =
-            tool_func!(async |_args: Value, _runenv: &dyn RunEnv| -> Value { Value::bool(true) });
+        let runenv = local_handle().await;
+        let f = tool_func!(async |_args: Value, _runenv: Arc<RunEnvHandle>| -> Value {
+            Value::bool(true)
+        });
         let out = f
-            .call(Value::object_empty(), "call-1", &runenv)
+            .call(Value::object_empty(), "call-1", runenv)
             .next()
             .await
             .unwrap();
@@ -439,12 +444,12 @@ mod tests {
 
     #[tokio::test]
     async fn test_stream_value() {
-        let runenv = Local {};
+        let runenv = local_handle().await;
         let f = tool_func!(stream |_args: Value| -> Value {
             stream::iter(vec![Value::integer(1), Value::integer(2), Value::integer(3)])
         });
         let outputs: Vec<_> = f
-            .call(Value::object_empty(), "call-1", &runenv)
+            .call(Value::object_empty(), "call-1", runenv)
             .collect()
             .await;
         assert_eq!(outputs.len(), 3);
