+++ src/tool/impl/builtins/read.rs
@@ -98,7 +98,7 @@ pub fn get_read_tool_desc() -> ToolDesc {
 
 pub fn get_read_tool_func() -> ToolFunc {
     tool_func!(
-        async |args: Value, id: String, runenv: &dyn RunEnv| -> Message {
+        async |args: Value, id: String, runenv: Arc<RunEnvHandle>| -> Message {
             let Some(path_str) = args.pointer("/path").and_then(|v| v.as_str()) else {
                 return error_message(id, "missing required parameter: path", "validation");
             };
@@ -173,7 +173,7 @@ mod tests {
     use futures::StreamExt;
 
     use super::*;
-    use crate::{datatype::Value, runenv::Local, to_value, tool::ToolProvider};
+    use crate::{datatype::Value, runenv::RunEnv, to_value, tool::ToolProvider};
 
     fn provider() -> ToolProvider {
         let mut p = ToolProvider::new();
@@ -185,7 +185,8 @@ mod tests {
         let provider = provider();
         let funcs = provider.provide(&[get_read_tool_desc()]).unwrap();
         let f = funcs.get("read").unwrap();
-        f.call(args, "1", &Local {}).next().await.unwrap().message
+        let runenv = RunEnv::local().get().await.unwrap();
+        f.call(args, "1", runenv).next().await.unwrap().message
     }
 
     #[tokio::test]
