+++ src/tool/impl/builtins/write.rs
@@ -29,7 +29,7 @@ pub fn get_write_tool_desc() -> ToolDesc {
 }
 
 pub fn get_write_tool_func() -> ToolFunc {
-    tool_func!(async |args: Value, runenv: &dyn RunEnv| -> Value {
+    tool_func!(async |args: Value, runenv: Arc<RunEnvHandle>| -> Value {
         let Some(path) = args.pointer("/path").and_then(|v| v.as_str()) else {
             return crate::to_value!({
                 "error": "missing required parameter: path",
@@ -61,7 +61,7 @@ mod tests {
     use futures::StreamExt;
 
     use super::*;
-    use crate::{datatype::Value, message::Message, runenv::Local, to_value, tool::ToolProvider};
+    use crate::{datatype::Value, message::Message, runenv::RunEnv, to_value, tool::ToolProvider};
 
     fn provider() -> ToolProvider {
         let mut p = ToolProvider::new();
@@ -73,7 +73,8 @@ mod tests {
         let provider = provider();
         let funcs = provider.provide(&[get_write_tool_desc()]).unwrap();
         let f = funcs.get("write").unwrap();
-        f.call(args, "1", &Local {}).next().await.unwrap().message
+        let runenv = RunEnv::local().get().await.unwrap();
+        f.call(args, "1", runenv).next().await.unwrap().message
     }
 
     #[tokio::test]
